requirements.txt
asgiref==3.6.0
bleach==5.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
commonmark==0.9.1
Django==4.1.5
django-environ==0.9.0
django-taggit==3.1.0
docutils==0.19
idna==3.4
importlib-metadata==6.0.0
jaraco.classes==3.2.3
jnunes-commons==0.3.13
keyring==23.13.1
more-itertools==9.0.0
pkginfo==1.9.6
Pygments==2.14.0
pywin32-ctypes==0.2.0
readme-renderer==37.3
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.0.1
six==1.16.0
sqlparse==0.4.3
twine==4.0.2
tzdata==2022.7
urllib3==1.26.13
webencodings==0.5.1
zipp==3.11.0
